move celebration to dedicated directory
