+++ src/home/celebration/NftCelebration.test.tsx
